patterns/01-full-load/requirements.txt
psycopg2-binary==2.9.9
PyYAML==6.0.2
pytest==8.3.3
